Update package.test.json
Creating scripts For The Build Process

For all tasks, in order to allow for online testing, please make changes as mentioned in package.json, and then copy the modified file over to package.test.json. Online tests will examine package.test.json.

In order to verify your work for this module, open your command-line application to the directory of your project and run npm run test:module2.

    1
    Create script that will transpile the code using Babel

Open package.json and go to the scripts section. Add a script named build with the value babel src -d build to transpile the source code using Babel.

Note:

    all we will be doing will be based on this file so keep it open and make changes to the object mapped to the scripts key.
    Ensure too for all tasks you remember to copy package.json to package.test.json (see above for why)

2
Add a script to clean up the build directory

There will be situations you'll need to cleanup the build directory. For that reason, add a script named build-clean to the scripts object and put rm -rf build as the value for that key.