Update package.test.json
postbuild-clean with a value of npm run build.---------------------------------------------------------------------------------------------------

Create script that will transpile the code using Babel

Open package.json and go to the scripts section. Add a script named build with the value babel src -d build to transpile the source code using Babel.

Note:

    all we will be doing will be based on this file so keep it open and make changes to the object mapped to the scripts key.
    Ensure too for all tasks you remember to copy package.json to package.test.json (see above for why)

2
Add a script to clean up the build directory

There will be situations you'll need to cleanup the build directory. For that reason, add a script named build-clean to the scripts object and put rm -rf build as the value for that key.
3
Add post hook for the build-clean script

Add post hook for the build-clean script by adding a new key named postbuild-clean with a value of npm run build.

The scripts object doesn't have a key named postbuild-clean.