start-dev >test.json
start-dev to the scripts object and give it the value nodemon --watch src --exec PORT=5000 npm start.

In order to verify your work for this module, open your command-line application to the directory of your project and run npm run test:module4.

    1

Add a start script to build and run the application
2
Add script to run the application in development mode

Add a key named start-dev to the scripts object and give it the value nodemon --watch src --exec PORT=5000 npm start.